Check people avatars in report email and test refactor
